Renamed GroupController and added controllers for roles and users
